Move ExprType to the AST
There are various places in the AST and other parts of the codebase
where we need ExprTypes, not VarTypes.  Start such changes by first
moving where the type is defiend.
